+++ src/components/content/medication-drawer.tsx
@@ -5,10 +5,8 @@ import { DataList, entryFromArray } from "../core/data-list";
 import type { DrawerProps } from "../core/drawer";
 import { Drawer } from "../core/drawer";
 import { MedicationHistory } from "./medication-history";
-import { useEffect, useState } from "react";
-import { useMedicationHistory } from "@/fhir/medications";
-import { map, get, filter, last, pipe, sortBy, propEq } from "lodash/fp";
-import { MedicationModel } from "@/fhir/models";
+import { useLastPrescriber } from "@/fhir/medications";
+import { Loading } from "@/components/core/loading";
 
 export type MedicationDrawerProps = {
   medication?: MedicationStatementModel;
@@ -37,38 +35,11 @@ export const MedicationDrawer = ({
   medication,
   ...drawerProps
 }: MedicationDrawerProps) => {
-  const [lastPrescriber, setLastPrescriber] = useState<string | undefined>();
+  const { lastPrescriber, isLoading } = useLastPrescriber(medication?.resource);
   const data = getDataEntriesFromMedicationStatement(
     medication,
     lastPrescriber
   );
-  const historyQuery = useMedicationHistory(medication?.resource);
-
-  useEffect(() => {
-    const { includedResources = {}, medications = [] } =
-      historyQuery.data || {};
-
-    if (lastPrescriber === undefined && medications.length) {
-      const prescriber = pipe(
-        // 1. Get underlying resources from the medication models.
-        map(get("resource")),
-        // 2. Throw away any resources that are not MedicationRequests.
-        filter(propEq("resourceType", "MedicationRequest")),
-        // 3. Sort by the authored on date.
-        sortBy((m) => Date.parse(m.authoredOn)),
-        // 4. If there are med dispense records, make them models.
-        map((mr) => new MedicationModel(mr, includedResources)),
-        // 5. Take the last (latest) from our filtered list.
-        last,
-        // 6. Get the prescriber from the medication model.
-        get("prescriber")
-      )(medications);
-
-      // Fall back to a string, so we don't try to find prescriber again.
-      setLastPrescriber(prescriber || "");
-    }
-  }, [lastPrescriber, historyQuery.data, setLastPrescriber]);
-
   return (
     <Drawer title="Medication Details" {...drawerProps}>
       <Drawer.Body>
@@ -78,7 +49,8 @@ export const MedicationDrawer = ({
               {medication?.display}
             </h3>
           </div>
-          <DataList title="Summary" data={data} />
+          {isLoading && <Loading />}
+          {!isLoading && <DataList title="Summary" data={data} />}
           {medication && <MedicationHistory medication={medication} />}
         </div>
       </Drawer.Body>
